money_exchange_app/requirements.txt
coverage==7.0.1
Flask==2.2.2
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
